+++ src/main/java/dev/fpsaraiva/libraryapi/api/exceptions/ApiErrors.java
@@ -2,6 +2,7 @@ package dev.fpsaraiva.libraryapi.api.exceptions;
 
 import dev.fpsaraiva.libraryapi.exception.BusinessException;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -18,6 +19,10 @@ public class ApiErrors {
 
     public ApiErrors(BusinessException ex) {
         this.errors = Arrays.asList(ex.getMessage());
+    }
+
+    public ApiErrors(ResponseStatusException ex) {
+        this.errors = Arrays.asList(ex.getReason());
     }
 
     public List<String> getErrors() {
