+++ src/test/java/dev/fpsaraiva/libraryapi/api/resource/LoanControllerTest.java
@@ -85,7 +85,7 @@ public class LoanControllerTest {
         mvc.perform(request)
                 .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("errors", Matchers.hasSize(1)))
-                .andExpect(jsonPath("errors[0]").value("Book not found for informed isbn"));
+                .andExpect(jsonPath("errors[0]").value("Book not found for informed isbn."));
 
     }
 }
